AG-9015 Move errorBar code into errorBar.ts
The implementation of error bars needs to be decoupled from the
CartesianSeries as much as possible.

The only thing that the error bar scene node need to know for now is the
the data points and X and Y scales (to convert data points into pixel
points). These variables are provided as parameters of the
ErrorBar.createNodeData method.