AG-9015 Add placeholder ErrorBarModule
This enterprise module doesn't do anything, but it provides the basic
framework to eventually move the errorBars.ts implementation from the
community package into the enterprise package.

The model used in the community package closely matches the connection
that the community axes shares with the enterprise crosshair. The Series
class now has utility to add & remove modules to it.

The Series class uses a new general purpose ModuleMap class to implement
this. The moduleMap.ts file has to same implementation axis.ts
(addModule, removeModule, isModuleEnabled, createAxisContext), but has
been generalised.

axis.ts is left unchanged for now, but eventually this will need to be
refactored to use the ModuleMap type like series.ts.