AG-9015 Set the yLowerKey and yUpperKey properties
There was a bug in ModuleMap where the class would set the optionKey
property on itself rather than the instance that uses the ModuleMap.
ModuleMap now sets the property of the 'parent' instance. This matches
the behaviour of the abstract class Axis (which will eventually need
to be refactored to use the ModuleMap class).

The yLowerKey and yUpperKey properties are now part of the ErrorBars
class (instead of a separate class). This matches the implementation of
the crosshairs module for consistency.