AG-9017 Add styling options to error bars
This adds the following options:

- visible (default: true)
- stroke (default: 'black')
- strokeWidth (default: 1)
- strokeOpacity (default: 1)

These options are duplicated for both the whiskers and caps (so that
these components can be customised individually). Cap inherit the
defaults from the whisky (e.g. setting whisker.stroke = 'red' will
also set cap.stroke = 'red')

The rendering of caps is not implemented yet.