AG-9015 Hide error bars when series is hidden
This generalises the CartesianSeries.dataModelListeners property, which
has been moved / renamed to Series.seriesListeners. The property is kept
private so that the base class Series can check that the event types of
the handlers and dispatchers match. Use Series.dispatch to dispatch
events to listeners.

This is currently only tested & used by errorBar.ts.

Note: For legacy reasons, Series extends Observable. This means it now
has both addEventListener (from Observable) and addListener (from
Listeners). The former is preferred for new code.