AG-9015 Extend the domain of line series to fit error bars
This adds a next series event 'processData-prerequest'. With this event,
the module can add, remove or modify any of the properties of the data
model before it is processed (in this case, the error bars adds the
lower & upper keys of the X & Y axes).

Only tested on line series. Probably won't work on other series.