AG-9014 Rename Series.getDomain to getSeriesDomain
Also add a new (non-abstract) Series.getDomain. This new method computes
the local union of the series plus any extended domain defined by series
option modules (error bars lower/upper data point for example). This new
getDomain is not intended to be overriden, but there does not seem to be
any way to ensure this.

Explicitly use the 'override' keywords when override the abstract
getSeriesDomain method.