AG-9014 Position errorbar using the band-scale in bar-series
This significantly refactors the errorBar.ts file. The ErrorBars class
now uses an abstract CartesianInterface class which behaves almost
exactly the same for line, scatter and bar series (but with some
overrides for bar-series). This cuts back on the number of if-checks
that the ErrorBars class need (because it doesn't really care about the
underlying series type).

With this change, the errorbars in bar-series are centered on the top
left corner of the bars for each series. This is still wrong (it should
be in the top center of the bar), but this is good progress.