AG-2017 Implement errorbar cap length configuration
This adds the `ErrorBoundSeriesNodeDatum` type, which is extended by the
`BarNodeDatum`, `LineNodeDatum` and `ScatterNodeDatum` types. The
errorBar.ts file has some type guards to enforce this (this is
implemented in the `toErrorBoundCartesianSeries` function).

Various tests have been added to errorBar.test.ts to check that the cap
length is rendered correctly.