Use series options in error checking
This prevents the need of exporting BarSeries, LineSeries and
ScatterSeries.

This losses the type safety check that the TDatum type of CartesianSeries
extends the ErrorBoundSeriesNodeDatum type. This introduces two risks:

1.  If the BarSeries, LineSeries, ScatterSeries are changed such that the
    TDatum type no longer extends the ErrorBoundSeriesNodeDatum, then
    this will cause a bug regression.

2.  If a new series types is added/modified with the intent of supporting
    error bars, then there will be runtime errors if the series does not
    respect the assumptions of errorBar.ts

The first risk is acceptable because our tests should catch this.

The second risk is acceptable because the developer writing this should
notice the problem.

Note: the AgErrorBarSupportedSeriesTypes constant is required because the
errorBar.ts file needs a string[] to perform its checks. Converting a
string[] to a union is trivial, but converting a union to a string[] is
more complicated. See:
https://github.com/microsoft/TypeScript/issues/13298#issuecomment-820634611