AG-8238 Listen for highlight events in errorBar.ts
Keeping the ErrorBarNode instance in the Series.contentGroup breaks the
rendering when an item is highlighted.

Keeping the ErrorBarNode instance in the Series.highlightGroup fixes this
rendering bug when an item is highlighted but breaks visibility toggling.

Therefore, we need to move the ErrorBarNode back and forth between the
contentGroup and highlightGroup when the highlighting changes.