AG-9272 Remove ErrorBoundSeriesNodeDatum.index
As a substitute, search sequentially for the matching `ErrorBarNode`
instance in the new method `ErrorBar.restyleHightlightChange`. This isn't
a good solution in terms of performance. However, the assumption right
now is that the typical use case for error bars includes few data points
(because the chart will get cluttered very quickly if there are many
data points with error bars). This node lookup can be revisited if
performance becomes an issue.

Also separate the style and translation updates in `ErrorBarNode`. This
allows use to specify whether and major or minor scene change occurs.