AG-9017 Only assign style properties to Paths
The result from formatters is user data and can include any arbitrary set
of key-values. Object.assign copies over all key-values from the source
to the target.

Users could potentionally exploit this to overwrite properties that aren't
intended to be modified. This is unsafe, so only assign an explicit set of
key-values.