AG-9272 Implement pickNode in errorBar.ts
After some discussion, we decided not to mutate the InteractionEvent.
Instead, errorBar.ts just implements hit-testing for exact, nearest
and mainAxisFirst pick modes.

The CartesianSeries class also loops through the modules to determine
which node is the nearest. With this, the ErrorBars class no longer
needs to listen for input events because that's handled by the Chart
class.