AG-9715 - Fix JS not working on pages with tabs after build
The problem has to do with the Astro Markdoc integration not running JS during build time if the html from `Astro.slots.render('default')` is not present on the page. This could be that it errors, or the Astro island is missing something and fails silently. The slot content can be parsed and data can be extracted from it during build time, but if the generated html is not on the page, the JS on that page doesn't work eg, can't expand/collapse menus or change tabs.

Basically in `TabsWithHtmlChildren.astro`:

```
---
const html = await Astro.slots.render('default');

// Can process `html` during build time, but need it to be on the page as well
const tabsData = extractTabsData(html);
---

<!-- Need this on the page -->
<Fragment set:html={html} />
```

Also renaming component back to `TabsWithHtmlChildren.tsx` as that is more accurate now.