AG-9715 - Add fix to Add documentation tabs to right hand side nav + scroll highlights (#453) (#472)
* AG-9715 - Add documentation tabs to right hand side nav + scroll highlights (#453)

* AG-3390 - Use `flatMap` instead of `map().flat()`

* AG-9715 - Add scroll highlights for right hand side nav

* AG-9715 - Add tab labels to right side menu and allow them to change tabs

* AG-9715 - Move tab data generation to astro build time

Allows it to load before client side code is run

Also move `DOCS_TAB_ITEM_ID_PREFIX` constant to top level, so that TabsWithData is generic

* AG-9715 - Update hash when tab is selected

* AG-9715 - Highlight right side menu tab heading on page load

* AG-9715 - Listen to hash changes to update side menu selection

* AG-9715 - Move side menu nav scrolling into react hook

* AG-9715 - Fix JS not working on pages with tabs after build

The problem has to do with the Astro Markdoc integration not running JS during build time if the html from `Astro.slots.render('default')` is not present on the page. This could be that it errors, or the Astro island is missing something and fails silently. The slot content can be parsed and data can be extracted from it during build time, but if the generated html is not on the page, the JS on that page doesn't work eg, can't expand/collapse menus or change tabs.

Basically in `TabsWithHtmlChildren.astro`:

```
---
const html = await Astro.slots.render('default');

// Can process `html` during build time, but need it to be on the page as well
const tabsData = extractTabsData(html);
---

<!-- Need this on the page -->
<Fragment set:html={html} />
```

Also renaming component back to `TabsWithHtmlChildren.tsx` as that is more accurate now.