AG-9715 - Rewrite tabs as astro component
The react component was getting server side and client side data mixed up and causing the page not to load. It might be that it's a complicated use case that involves parsing markdoc into html, extracting tab data from the markdoc data and then rendering the tabs as react. Removing the react part of it simplifies the tech stack and allows Astro to optimize the server rendering part, while the client side can be handled by a standard Web Component. In theory, because most of the logic is a Web Component, it can be ported to other future tech stacks.