AG-9715 - Make example runner `client:only`
There were issues with example runner JS not loading and hence not loading the examples in the iframe. Making the component `client:only` means all of it runs on the client side, and fixes this issue.

To make sure there is no layout shift during loading, the `minHeight` of the example is also moved to the container level.