AG-9963 Mute duplicate warnings when axis position is valid.
The bug here that the position always defaults to 'left'. This that if
the axes configuration something like 0: 'left' and 1: 'bottomasdf'
then this will default to 0: 'left' and 1: 'left'.

This is what the causes the duplicate warning, because the library
cannot determine which is the X axis.

To get around this, we'll change the default based on any valid
positions that have been identified.