AG-10326 Use an empty colorRange as the default
The colorRange property need to be set to a PropertiesArray so it can
validate the color and stop properties. However, the "default color range"
uses an interpolation of the background Fill and series fill, so it
cannot be calculated at compile-time.

Therefore, we'll use an empty colorRange to mean "interpolate". As a
consequence though, no warning message is shown when colorRange is
explicitly set to [].