AG-10323 Fix tooltips on mobile devices
The bug is caused because because we disabled the pointer whenever a drag
event starts (i.e. when 'mousedown' or 'touchstart' is received). Now we
only disable the pointer when the drag actually starts moving.

This also fixes a bug regression where clicking on a node with the mouse
(on desktop) hides the tooltip and highlighting.