AG-10657 Fix warnings about excessive updates
The 'series' region from the RegionManager is the first listener to
receive 'wheel' events. The Zoom module consumes 'wheel' events which
prevents the Chart class from receiving these events and removing the
highlight.

Changing the Chart to listenfor 'wheel' events in the 'series' region
ensures that it receives these events before the Zoom module.

Without this change, the following warning would be printing when
scrolling on a zoomable chart:

```
AG Charts - exceeded the maximum number of simultaneous updates (4), discarding changes and rendering
```