AG-10923 Add a new 'zoom-pan-start' ZoomManager event.
This will keep the ordering of events & listener more predictable, by
ensuring that the chart clears the highlight when the zoom / navigator
modules start dragging.

We still need to keep the 'drag' listener in chart.ts for now though.
Without it, the user can keep highlight active by dragging out of the
chart's rect.