Rename 'annotation' layer to 'addon'
This layer/group was intented to be used by error-bars. The error-bars
adds data & nodes to the series itself, so the name 'addon' is more
suitable. It will also avoid confusion, because the term 'annotations'
also refers to a new feature that's currently in development.

In fact, it seems as though some confusion has already happened,
because we are apparently using this layer for both error-bars and
annotations. This is probably an unintentional oversight.