AG-11192 Ignore keyboard events in Chart.handlePointer
When two series have datum that overlap one another, the keyboard
navigates through which ever datum is rendered on top like the mouse.
We don't want that.

Example:
{
  data: [
    { x: 0, foo: 1, bar: 3 },
    { x: 1, foo: 2, bar: 2 },
    { x: 2, foo: 3, bar: 1 },
  ],
  series: [
    { xKey: 'x', yKey: 'foo' },
    { xKey: 'x', yKey: 'bar' },
  ]
}

Lines 'foo' and 'bar' intersect at (x,y) = (1,2), with 'foo' rendered
on top.

When navigator left/right on 'foo' the tooltip will show:
 "foo: 1", "foo: 2", "foo: 3"

When navigator left/right on 'bar' the tooltip will show:
 "bar: 3", "foo: 2", "bar: 1" (undesirable behaviour)

This change makes the tooltip show this instead:
 "bar: 3", "bar: 2", "bar: 1"