AG-11190 Skip regions that the keyboard cannot interact with
The captions might hidden for example, so we only want to focus on
captions that will do something when they receive a 'hover'-like event
from the keyboard.

This is untested with resizing. Currently the RegionManager just assumes
that the currentTabIndex is interactable, we might not necessarily be the
case. For example: if the chart grows in size, then a caption that was
truncated could end up fitting on the screen and not show tooltip.