Initial InteractionState as 'Animation'
The listener in chart.ts must ignore interactions during animation.
However, the animation's initialisation takes a while to complete, which
means that the click events in chart.ts can be fired (and cancel the
animation) if the user clicks very quickly at the start.

Using InteractionState.Animation as the initial state solves this;
the AnimationManager will automatically pop this state if animations are
disabled.