AG-3390 - Add prettier ignore file for local dev
There are currently issues with the alignment of prettier/eslint rules on grid and charts, so the shared folder sometimes has formatting issues that break CI for charts. This meant that we needed to add the shared folder into prettier ignore to prevent false negatives on CI. However, this means that on local dev, we lose prettier functionality. Add a copy of `.prettierignore` that excludes the shared folder, into VS code settings, so that it can be prettified on local development.

This file and VS code setting can be removed when there is alignment between grid/charts with the prettier/eslint setup.