AG-10486 Refactor contextmenu event handling
Move the input handling into ContextMenuRegistry (which may be renamed to
ContextMenuManager at some point in the near future). Instead, just
dispatch abstract "context" events.

Eventually, the legend and series contextmenu listeners will specify the
parametered like `pickedNode` or `pickedLegendItem` as part of the
context event.

This will significantly reduce the complexity of the ContextMenu module,
which is responsible for too much right now.