AG-11676 Change how Zoom double click works
Now, the Chart listener is fired first and it consumes the event if the
double click is on a node.

This fixes a bug with `listeners.doubleClick` option, and it prevents the
need to rely on HighlightManager.getActivePicked() in the Zoom module.
Which in turn will also this to be moved to the ContextMenuEvent.context
property like with the legend.