AG-11917 AG-11899 Set pointer-events: none on overlay
There is a recent bug regression where the KeyNavManager is not correctly
receiving keyboard events.

This happening because we’ve swapped the DOM ordering of the canvas and
overlay elements. This was done because we need keep the correct
tab-ordering (the canvas, i.e. series, is the first region when pressing
TAB).

Setting `pointer-events: none` on the .ag-charts-canvas-overlay element
allow mouse event to pass-through and hit the <canvas> element, which
ensures that it also receives keyboard input.