AG-11194 Bugfix: sometimes toolbar buttons could have two tabIndex=0
Screenreaders have built-in features and commands that keep the user more
freedom on this elements to focus on.

For example, in VoiceOver the user can press Ctrl+Opt+Shift Up to leave
the toolbar button (and enter into the toolbar parent itself), and then
focus on the first item by pressing Ctrl+Opt+Shift Down to enter the
toolbar again.

Therefore, only relying on Left / Right key presses can result in two
tabIndex=0 values in the toolbar, which is undesired.