AG-11950 Ensure that the legend markers remain aligned
In combo charts, the user can have a bar series and line series with
showSeriesStroke: true. This will causes the legend markers to be
unaligned on vertical legends, because the bar legend-marker has a
shorter width than the line legend-marker.

This commit ensures that legend-markers all have the same width.