AG-12369 Detect mouse clicks downside menus and close them
Putting in logic in keynavUtil.ts is a bit counter-intuitive, because it
was meant to handle keyboard interactions and not mouse interactions. But
as the same, putting this logic in initMenuKeyNav is the best way to
ensure consistency with context menus and financial charts popovers.

We may want to rename this file / functions at some point to make it
clearer that both mouse and keyboard event handlers are initialised.