AG-12369 Refactor initMenuKeyNav params and return-type
It now takes a `device` parameter (mouse or keyboard). This parameter is
used to determine what the element to focus on when opening the menu, and
what element to focus on when closing the menu (if applicable).

The method now also return the { close: ()=>void } object which can be
called in `Popover.hide()` and `ContextMenu.hide()` to ensure that focus
is correct reset when clicking a menu item with the keyboard.