AG-NONE Ignore debug error log of thumbnail executor
The exception thrown by the generateThumbnail contains far more useful
information than a new error thrown by the executor.ts

In this instance, the console errors indicate to me that an exception is
being thrown by the constructor of the DOMManager. That's very helpful,
because that only one recent change there that could be responsible.