AG-12894 Fix ContextMenu cleanup logic
The MenuCloser.close() function relies on the state of this.element.
Therefore, we need to ensure that we close the existing contextmenu
before updating the content on this.element.

Without this change, an error will be thrown by the JS engine because
we are calling removeElement(elem) where elem isn't a child.