AG-13041 Don't show focus after deleting annotations with keyboard
This fixes a regression caused by the removal of the "mouseBlur" feature in
KeyNavManager.

This is sort of an anti-pattern. Showing the focus after pressing Backspace or
Delete on the keyboard is the "correct" behaviour. However, the correct
behaviour in this scenario feels flimsy.

Rather than restoring the "mouseBlur" feature, we instead create and focus on a
temporary so-called "focus mask" which is added above the canvas. The focus mask
does not respond to pointer events, and puts the canvas-proxy back into focus
whenever any key is pressed. The focus mask is removed once blurred. This mimics
the "mouseBlur" functionality that the use to have.