AG-13142 Move FocusIndicator out of ChartContext
The FocusIndicator lives in the ChartContext because is what historically used
for all controls (e.g. navigator & legend). But it's only used by the
series-area right now, so it makes for sense for the SerieaAreaManager to own
this instance.

It's still used in KeyNavManager too though. Therefore, set this property in the
SeriesAreaManager as a temporary workaround for now.