AG-13357 Fix more eslint error
1 @typescript-eslint/no-floating-promises
2 prefer-const

The `sonarjs/no-redundant-optional` error in object.ts is incorrect. The
`undefined` type and `?` specifier are NOT redundant. They do not mean the same
thing. The `?` means that this properties does not need to be specified, and it
also implicitly makes the parameter a type union with `| undefined`
included (which in this case is `undefined | undefined` which is just
`undefined`). Whereas the `undefined` type just declares that the `set` property
must be of type `undefined` in this overload, that's it.

Removing the `?` will cause this code compilation to fail:

    defineSetterGetter(myObj, 'myProp', { get: () => myProp } );

The line above would need to be changed to:

    defineSetterGetter(myObj, 'myProp', { set: undefined, get: () => myProp } );