AG-13489 Fix e2e test 'zoom > navigator'
We need offsets relative to the HTMLElement that we are listening to. The
RegionManager has canvasXY for that. These offsets are useful, so add a
generalised version of these to MouseWidgetEvent: currentXY, which are offsets
relative to the sourceEvent.currentTarget (which is the widget that is listening
for events).

Also, add listeners to the containerWidget. This is because the zoom toolbar is
not a descendant of .ag-charts-series-area, so a 'mouseleave' is fired whenever
the pointer hovers over a zoom button, which hides the toolbar. We don't want
that. However, the zoom toolbar is a descendant of .ag-charts-canvas-container
so we can add a listener there instead. This also has the added affect of
showing the zoom toolbar when hovering over the axes (when onHover's bounds
hit-testing logic is true), which is the expected behaviour in production.