AG-13798 Fix 'wheel' zoom events in Safari
This might be because the `passive` defaults in Safari is different to other
browsers. Regardless, we're adding `'wheel'` event listeners to the
.ag-charts-series-area element which has the CSS value `pointer-events:
none`. This is sort of asking for trouble.

Making the .ag-charts-series-area have `pointer-event: auto` works on all
browsers (Safari, Firefox, Chrome), although we need to set the
.ag-charts-series-area's tabindex to -1 and focus on the .ag-charts-swapchain
child when the series area is clicked on.