AG-8876 Allocate expandedItems without intermediate result
Note, the ContextMenu.expandItem member property is not required. This variable
is only used as an intermediate to allocate & initialise the context menu UI
elements.

Also, the initial algorithm expands all context menu items, and then filters out
hidden items (based on the `showOn` values) in a second pass. This is unnecessary.