AG-15830 Use a stricter type-rule for datumIndex
The exception in this bug is being thrown by
`LinearGaugeSeries.getTooltipContent()` because the datumIndex param (value:
`undefined`, type: `unknown`') is silently being converted into a non-nullable
LinearGaugeNodeDatumIndex object type, and then the `.type` is being
dereferences on an undefined value.

The naive fix to just add a check in LinearGaugeSeries.getTooltipContent for
null `datumIndex`, but that's too easy. The TypeScript compiler should catch
this error at compile-time, and enforce null-checks where applicable.

This patch breaks compilation. This is expected, the compiler is now find places
there the nullish type rules are not explicitly specified.