AG-15782 Restrict callWithContext to single-arg functions
The callWithContext function is only ever called with one argument, and wouldn't
work when being called with multiple arguments anyway. See function
`maybeSetContext` in `callbackCache.ts`; it deferences `params[0]`.

This change cleans up & simplifies the code, but only help to workaround an
error when trying to generalise the `getStyle` implementation of RadarAreaSeries
and RadarLineSeries.